[WIP] Move parse_indexes() from list.c to liblzma
The first steps toward moving this functionality to `liblzma` is to
decouple it from I/O. As per the TODO comment on `parse_indexes()`
annotated documentation, we can accomplish this by passing a callback to
perform the actual I/O.

This commit introduces the following refactorings:

- Make `parse_indexes()` take the file size as a parameter rather than
getting it from the `file_pair *`.

- Make `parse_indexes()` the file name as a parameter rather than
getting it from the `file_pair *` (this is only used for error reporting
reasons).

- Make `parse_indexes()` take a callback function to perform the I/O.

- Make `parse_indexes()` take an opaque pointer rather than a `file_pair
*`, so that the client can pass any kind of file representation he
wants, and handle accordingly on the callback.

- Replace `io_buf *` by `uint8_t *`.

- Create and use a new `lzma_file_info` struct rathern than the
`xz`-specific `xz_file_info`.

Signed-off-by: Juan Cruz Viotti <jviottidc@gmail.com>